function-app/requirements.txt
azure-functions
azure-identity
openai>=1.30.0
